+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/ConnectedApp/ConnectedAppContainer.tsx
@@ -16,6 +16,7 @@ import {useFetchConnectedAppMonitoringSettings} from '../../hooks/use-fetch-conn
 import {MonitoringSettings} from '../../../model/Apps/monitoring-settings';
 import {ConnectedAppErrorMonitoring} from './ErrorMonitoring/ConnectedAppErrorMonitoring';
 import {DeveloperModeTag} from '../DeveloperModeTag';
+import isGrantedOnProduct from '../../is-granted-on-product';
 
 type Props = {
     connectedApp: ConnectedApp;
@@ -163,6 +164,8 @@ export const ConnectedAppContainer: FC<Props> = ({connectedApp}) => {
 
     const tag = connectedApp.is_test_app ? <DeveloperModeTag /> : null;
 
+    const isAtLeastGrantedToViewProducts = isGrantedOnProduct(connectedApp, 'view');
+
     return (
         <>
             <PageHeader
@@ -199,7 +202,7 @@ export const ConnectedAppContainer: FC<Props> = ({connectedApp}) => {
                     >
                         {translate('akeneo_connectivity.connection.connect.connected_apps.edit.tabs.settings')}
                     </TabBar.Tab>
-                    {null !== providers && providers.length > 0 && (
+                    {null !== providers && providers.length > 0 && isAtLeastGrantedToViewProducts && (
                         <TabBar.Tab
                             isActive={isCurrent(permissionsTabName)}
                             onClick={() => {
@@ -233,11 +236,12 @@ export const ConnectedAppContainer: FC<Props> = ({connectedApp}) => {
                     />
                 )}
 
-                {isCurrent(permissionsTabName) && null !== providers && (
+                {isCurrent(permissionsTabName) && null !== providers && isAtLeastGrantedToViewProducts && (
                     <ConnectedAppPermissions
                         providers={providers}
                         setProviderPermissions={handleSetProviderPermissions}
                         permissions={permissions}
+                        onlyDisplayViewPermissions={!isGrantedOnProduct(connectedApp, 'edit')}
                     />
                 )}