+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/UseCase/Validator/ListProductsQueryValidator.php
@@ -56,6 +56,8 @@ final class ListProductsQueryValidator
     /** @var ValidateGrantedLocalesInterface */
     private $validateGrantedLocales;
 
+    private ValidateIdentifiersLimit $validateIdentifiersLimit;
+
     public function __construct(
         ValidateAttributes $validateAttributes,
         ValidateChannel $validateChannel,
@@ -69,7 +71,8 @@ final class ListProductsQueryValidator
         ValidateGrantedCategoriesInterface $validateGrantedCategories,
         ValidateGrantedPropertiesInterface $validateGrantedProperties,
         ValidateGrantedAttributesInterface $validateGrantedAttributes,
-        ValidateGrantedLocalesInterface $validateGrantedLocales
+        ValidateGrantedLocalesInterface $validateGrantedLocales,
+        ValidateIdentifiersLimit $validateIdentifiersLimit
     ) {
         $this->validateAttributes = $validateAttributes;
         $this->validateChannel = $validateChannel;
@@ -84,6 +87,7 @@ final class ListProductsQueryValidator
         $this->validateGrantedProperties = $validateGrantedProperties;
         $this->validateGrantedAttributes = $validateGrantedAttributes;
         $this->validateGrantedLocales = $validateGrantedLocales;
+        $this->validateIdentifiersLimit = $validateIdentifiersLimit;
     }
 
     /**
@@ -107,6 +111,7 @@ final class ListProductsQueryValidator
         $this->validateCategories->validate($query->search);
         $this->validateGrantedCategories->validate($query->search);
         $this->validateProperties->validate($query->search);
+        $this->validateIdentifiersLimit->validate($query->search);
         $this->validateGrantedProperties->validate($query->search);
         $this->validateSearchLocales->validate($query->search, $query->searchLocaleCode);
         $this->validateGrantedLocales->validateForLocaleCodes($query->localeCodes);